clean up from failing tests
