+++ app/src/organisms/Devices/ProtocolRun/__tests__/LabwareInfoOverlay.test.tsx
@@ -10,10 +10,10 @@ import fixture_tiprack_300_ul from '@opentrons/shared-data/labware/fixtures/2/fi
 import { nestedTextMatcher, renderWithProviders } from '@opentrons/components'
 import { i18n } from '../../../../i18n'
 import { useCurrentRun } from '../../../ProtocolUpload/hooks'
-import { useProtocolDetailsForRun } from '../../hooks'
 import { getLabwareLocation } from '../utils/getLabwareLocation'
 import { LabwareInfoOverlay } from '../LabwareInfoOverlay'
 import { getLabwareDefinitionUri } from '../utils/getLabwareDefinitionUri'
+import { useLabwareOffsetForLabware } from '../useLabwareOffsetForLabware'
 
 jest.mock('../../../ProtocolUpload/hooks')
 jest.mock('../utils/getLabwareLocation')
@@ -45,8 +45,8 @@ const mockGetLabwareDisplayName = getLabwareDisplayName as jest.MockedFunction<
 const mockUseCurrentRun = useCurrentRun as jest.MockedFunction<
   typeof useCurrentRun
 >
-const mockUseProtocolDetailsForRun = useProtocolDetailsForRun as jest.MockedFunction<
-  typeof useProtocolDetailsForRun
+const mockUseLabwareOffsetForLabware= useLabwareOffsetForLabware as jest.MockedFunction<
+  typeof useLabwareOffsetForLabware
 >
 const mockGetLabwareLocation = getLabwareLocation as jest.MockedFunction<
   typeof getLabwareLocation
@@ -84,16 +84,15 @@ describe('LabwareInfoOverlay', () => {
       .calledWith(props.definition)
       .mockReturnValue('mock definition display name')
 
-    when(mockUseProtocolDetailsForRun)
-      .calledWith(MOCK_RUN_ID)
+    when(mockUseLabwareOffsetForLabware)
+      .calledWith(MOCK_RUN_ID, MOCK_LABWARE_ID)
       .mockReturnValue({
-        protocolData: {
-          commands: [],
-          labware,
-          labwareDefinitions,
-          modules: [],
-        },
-      } as any)
+        id: 'fake_offset_id' ,
+        createdAt: 'fake_timestamp',
+        definitionUri: 'fake_def_uri' ,
+        location: {slotName: MOCK_SLOT_NAME},
+        vector: MOCK_LABWARE_VECTOR
+      } )
 
     when(mockUseCurrentRun)
       .calledWith()
